requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
flask
pypdf
python-dotenv
ruff
setuptools
langchain-community
langchain_huggingface
pyngrok